chore: extract all spell lists to files
